Initial Commit
Initial commit. 2 half-ish days of work into the project. Reached a point where storing periodic .zips of the current state of the project was likely going to be insufficient for future progress. All textures (save for the arrow) are from Warr1024's NodeCore (check it out sometime), and as such are his property that I am borrowing.